Solved Issue #67
Now SaveAs has the expected behavior.

+++ sci_corpus/container.py
@@ -749,10 +749,11 @@ class ContainerDB():
         if path == '':
             path = self.path
             try:
-				self.__defaultpath = os.path.join(workspace,'backup.db')
-				if os.path.abspath(self.path) != os.path.abspath(self.__defaultpath):
-					copy2(path, os.path.abspath(self.__defaultpath))
-				os.remove(path)
+                self.__defaultpath = os.path.join(workspace,'backup.db')
+                if os.path.abspath(self.path) != os.path.abspath(self.__defaultpath):
+                    copy2(path, os.path.abspath(self.__defaultpath))
+                if os.path.exists(path):
+                    os.remove(path)
             except OSError as e:
                 print ("Error: %s - %s." % (e.filename, e.strerror))
             else:
